+++ InvestmentCalculator/Models/TotalOutputModel.cs
@@ -7,8 +7,8 @@ namespace InvestmentCalculator.Models
 {
     public class TotalOutputModel
     {
-        public double TotalSubscription { get; set; }
-        public double TotalInterest { get; set; }
-        public double TotalInvestment { get; set; }
+        public decimal TotalSubscription { get; set; }
+        public decimal TotalInterest { get; set; }
+        public decimal TotalInvestment { get; set; }
     }
 }
